requirements.txt
streamlit
newspaper3k
python-dotenv
langchain_groq
lxml_html_clean
langchain
